% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array3d.R
\name{array3d}
\alias{array3d}
\title{raster to nc 3d}
\usage{
array3d(u, z, lat, lon)
}
\arguments{
\item{u}{Array.}

\item{z}{Integer; Level (asusming third dimencion)}

\item{lat}{Numeric; latitude coordinate.}

\item{lon}{Numeric; longitude coordinate.}
}
\value{
raster
}
\description{
\code{\link{array3d}} Creates raster from nc path
}
