% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array4d.R
\name{array4d}
\alias{array4d}
\title{raster to nc 4d}
\usage{
array4d(u, z, lat, lon)
}
\arguments{
\item{u}{Array.}

\item{z}{Integer; Level (asusming third dimencion)}

\item{lat}{Numeric; latitude coordinate.}

\item{lon}{Numeric; longitude coordinate.}
}
\description{
\code{\link{array4d}} Creates raster from nc path
}
