% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cetesb.R
\docType{data}
\name{cetesb}
\alias{cetesb}
\title{Location some air quality stations CETESB}
\format{A SpatialPointsDataFrame with 23 stations
\describe{
  \item{estacion}{Station}
}}
\source{
\url{http://cetesb.sp.gov.br/ar/publicacoes-relatorios/}
}
\usage{
data(cetesb)
}
\description{
Location some air quality stations CETESB
}
\keyword{datasets}
