% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aao.R
\name{get_aao}
\alias{get_aao}
\title{Get Antartic Oscillation (AAO) data}
\usage{
get_aao()
}
\value{
data.frame
}
\description{
\code{\link{get_aao}} Downloads AAO data returning a data.frame
}
\note{
From http://www.cpc.noaa.gov/products/precip/CWlink/daily_ao_index/aao/aao.shtml:
The daily AAO index is constructed by projecting the daily (00Z)
 700mb height anomalies poleward of 20°S onto the loading pattern of the AAO.
 Please note that year-round monthly mean anomaly data
 have been used to obtain the loading pattern of the AAO.
 Since the AAO has the largest variability during the cold season,
 the loading pattern primarily captures characteristics of the cold season
 AAO pattern.
}
\examples{
{
print(head(get_aao()))
}
}
