% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_index.R
\name{get_all_index}
\alias{get_all_index}
\title{Get all index in long format}
\usage{
get_all_index(nao = TRUE, olr = TRUE, oni = TRUE, pdo = TRUE,
  pna = TRUE, soi = TRUE, sstoi = TRUE)
}
\arguments{
\item{nao}{Logica; inlude nao?}

\item{olr}{Logica; inlude olr?}

\item{oni}{Logica; inlude oni?}

\item{pdo}{Logica; inlude pdo?}

\item{pna}{Logica; inlude pna?}

\item{soi}{Logica; inlude soi?}

\item{sstoi}{Logica; inlude sstoi?}
}
\value{
data.frame
}
\description{
\code{\link{get_all_index}} Downloads all index returning a data.frame
}
\note{
From https://www.ncdc.noaa.gov/teleconnections/enso/indicators/
}
\examples{
\dontrun{
#do not run
a <- get_all_index()
head(a)
tail(a)
b <- get_all_index(olr = FALSE)
library(ggplot2)
library(cptcity)
ggplot(b, aes(x = Date, y = index, colour = index)) + geom_line()+
facet_wrap(~name, ncol = 2, scales = "free") + theme_bw() +
scale_colour_gradientn(colours = rev(cpt(find_cpt("cb_div_RdB")[2])), limit = c(-4.6, 4.6))
head(b)
tail(b)
}
}
