% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ao.R
\name{get_ao}
\alias{get_ao}
\title{Get Arctic Oscillation (AO) data}
\usage{
get_ao()
}
\value{
data.frame
}
\description{
\code{\link{get_ao}} Downloads AO data returning a data.frame
}
\note{
From https://www.ncdc.noaa.gov/teleconnections/ao/:
AO index is obtained by projecting the AO loading pattern to the daily anomaly
1000 millibar height field over 20N-90N latitude. The AO loading pattern has
been chosen as the first mode of EOF analysis using monthly mean 1000 millibar
height anomaly data from 1979 to 2000 over 20N-90N.
}
\examples{
{
print(head(get_ao()))
}
}
