% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nao.R
\name{get_nao}
\alias{get_nao}
\title{Get North Atlantic Oscillation (NAO) data}
\usage{
get_nao()
}
\value{
data.frame
}
\description{
\code{\link{get_nao}} Downloads NAO data returning a data.frame
}
\note{
From https://www.ncdc.noaa.gov/teleconnections/enso/indicators/nao/:

The North Atlantic Oscillation (NAO) index is based on the surface sea-level
pressure difference between the Subtropical (Azores) High and the Subpolar Low.
The positive phase of the NAO reflects below-normal heights and pressure across
the high latitudes of the North Atlantic and above-normal heights and pressure
over the central North Atlantic, the eastern United States and western Europe.
The negative phase reflects an opposite pattern of height and pressure anomalies
over these regions. Both phases of the NAO are associated with basin-wide changes
in the intensity and location of the North Atlantic jet stream and storm track,
and in large-scale modulations of the normal patterns of zonal and meridional
heat and moisture transport, which in turn results in changes in temperature and
precipitation patterns often extending from eastern North America to western and
central Europe.
}
\examples{
{
print(head(get_nao()))
}
}
