% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_olr.R
\name{get_olr}
\alias{get_olr}
\title{Get Outgoing long wave radiation (OLR) equator 160E-160W standarized data}
\usage{
get_olr()
}
\value{
data.frame
}
\description{
\code{\link{get_olr}} Downloads OLR data returning a data.frame
}
\note{
From https://www.ncdc.noaa.gov/teleconnections/enso/indicators/olr/

Outgoing Longwave Radiation (OLR) data at the top of the atmosphere are observed
from the Advanced Very High Resolution Radiometer (AVHRR) instrument aboard the
NOAA polar orbiting spacecraft. Data are centered across equatorial areas from
160E to 160W longitude. The raw data are converted into a standardized anomaly
index. Negative (Positive) OLR are indicative of enhanced (suppressed) convection
and hence more (less) cloud coverage typical of El Nino (La Nina) episodes.
More (Less) convective activity in the central and eastern equatorial Pacific
implies higher (lower), colder (warmer) cloud tops, which emit much less (more)
infrared radiation into space.
}
\examples{
\dontrun{
#dont run
}
}
