% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oni.R
\name{get_oni}
\alias{get_oni}
\title{Get Oceanic Nino Index (ONI) data}
\usage{
get_oni()
}
\value{
data.frame
}
\description{
\code{\link{get_ao}} Downloads ONI data returning a data.frame
}
\note{
From http://origin.cpc.ncep.noaa.gov/products/analysis_monitoring/ensostuff/ONI_v5.php:
Includes a threshold of +/- 0.5C for the Oceanic Nino Index (ONI)
3 month running mean of ERSST.v5 SST anomalies in the Nino 3.4 region (5N-5S, 120-170W),
}
\examples{
\dontrun{
#dont run
}
}
