% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pna.R
\name{get_pna}
\alias{get_pna}
\title{GetPacific Decadal Oscillation (pna) data}
\usage{
get_pna()
}
\value{
data.frame
}
\description{
\code{\link{get_pna}} Downloads pna data returning a data.frame
}
\note{
From https://www.ncdc.noaa.gov/teleconnections/pna/:

The Pacific-North America (PNA) pattern is one of the most prominent modes
of low-frequency variability in the Northern Hemisphere extratropics, appearing
in all months except June and July. The PNA pattern reflects a quadripole pattern
of 500 millibar height anomalies, with anomalies of similar sign located south
of the Aleutian Islands and over the southeastern United States. Anomalies with
sign opposite to the Aleutian center are located in the vicinity of Hawaii, and
over the intermountain region of North America (central Canada) during the winter
and fall.
}
\examples{
\dontrun{
#dont run
}
}
