% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soi.R
\name{get_soi}
\alias{get_soi}
\title{Get Outgoing long wave radiation (soi) equator 160E-160W standarized data}
\usage{
get_soi()
}
\value{
data.frame
}
\description{
\code{\link{get_soi}} Downloads soi data returning a data.frame
}
\note{
From https://www.ncdc.noaa.gov/teleconnections/enso/indicators/soi/

The Southern Oscillation Index (SOI) is a standardized index based on
the observed sea level pressure differences between Tahiti and Darwin,
Australia. The SOI is one measure of the large-scale fluctuations in air
pressure occurring between the western and eastern tropical Pacific (i.e.,
the state of the Southern Oscillation) during El Nino and La Nina episodes.
In general, smoothed time series of the SOI correspond very well with changes
in ocean temperatures across the eastern tropical Pacific. The negative phase
of the SOI represents below-normal air pressure at Tahiti and above-normal air
pressure at Darwin. Prolonged periods of negative (positive) SOI values coincide
with abnormally warm (cold) ocean waters across the eastern tropical Pacific
typical of El Nino (La Nina) episodes.
}
\examples{
\dontrun{
#dont run
}
}
