% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sstoi.R
\name{get_sstoi}
\alias{get_sstoi}
\title{Get Sea Surface Temperature Oceanic Index Data (sstoi)}
\usage{
get_sstoi(long = FALSE, only_anomaly = TRUE)
}
\arguments{
\item{long}{Logical; Return data.frame in long format?}

\item{only_anomaly}{Logical; only anomalies??}
}
\value{
data.frame
}
\description{
\code{\link{get_sstoi}} Downloads ONI data returning a data.frame
}
\note{
From https://www.ncdc.noaa.gov/teleconnections/enso/indicators/sst/
}
\examples{
{
print(head(get_sstoi(long = TRUE)))
}
}
