% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_regcm.R
\name{raster_regcm}
\alias{raster_regcm}
\title{Creates raster from nc path to RegCM}
\usage{
raster_regcm(nc, var, array = FALSE, verbose = TRUE)
}
\arguments{
\item{nc}{Character; Path to NetCDF.}

\item{var}{Character; name of variable. If you dont know, you can choose.}

\item{array}{Logical; return array or not.}

\item{verbose}{Logical; To show operations.}
}
\value{
raster
}
\description{
\code{\link{raster_nc}} Creates raster from nc path
}
\examples{
\dontrun{
#dont run
}
}
