% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_wrf.R
\name{raster_wrf}
\alias{raster_wrf}
\title{Creates raster from nc path}
\usage{
raster_wrf(nc, var, z, array = FALSE, verbose = TRUE)
}
\arguments{
\item{nc}{Character; Path to NetCDF.}

\item{var}{Character; name of variable. If you dont know, you can choose.}

\item{z}{Integer; Level (asusming third dimencion)}

\item{array}{Logical; return array or not.}

\item{verbose}{Logical; To show operations.}
}
\value{
raster
}
\description{
\code{\link{raster_nc}} Creates raster from nc path
}
\examples{
\dontrun{
#dont run
}
}
