% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratmos.R
\docType{package}
\name{ratmos}
\alias{ratmos}
\alias{ratmos-package}
\title{ratmos: a package to read gridded model outputs, convert them to raster and plot}
\description{
This package is designed to have few dependencies, raster, ncdf4 and sp..
The functions are designed so that, you only enter with the  path of the netcdf file
and then you can choose variables and levels interactivly
}
