% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqraster.R
\name{seqraster}
\alias{seqraster}
\title{Sequence of raster}
\usage{
seqraster(raster)
}
\arguments{
\item{raster}{raster}
}
\description{
\code{\link{seqraster}} creates a secuence of absolutes. For
instance, if the min of -0.3 and max 1.2, thje resulting sequence it would
be from -1.2 to 1.2 by 1/100.
}
