% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk.R
\name{tk}
\alias{tk}
\title{Temperature absolute Kelvin}
\usage{
tk(temperature)
}
\arguments{
\item{temperature}{Numeric temperature in celcius}
}
\description{
\code{\link{tk}} Calculates absolute temerature Kelvin
}
\examples{
\dontrun{
#dont run
}
}
\references{
Bolton, D. (1980). The computation of equivalent potential
temperature. Monthly weather review, 108(7), 1046-1053.
Equation 1
}
