% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tlcl.R
\name{Tncl}
\alias{Tncl}
\title{Temperature at the Lifting Condensation Level}
\usage{
Tncl(tk, rh)
}
\arguments{
\item{tk}{Numeric; temperature in kelvin}

\item{rh}{Numeric; relative humidity in percentage}
}
\description{
\code{\link{tk}} Calculates absolute temperature
}
\examples{
\dontrun{
#dont run
}
}
\references{
Bolton, D. (1980). The computation of equivalent potential
temperature. Monthly weather review, 108(7), 1046-1053.
Equation 22
}
