% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_shear.R
\name{wind_shear}
\alias{wind_shear}
\title{Calculates Wind Shear}
\usage{
wind_shear(u, v, zlev, at_zlev, zdiff, ascending = TRUE,
  dataframe = FALSE)
}
\arguments{
\item{u}{Numeric;}

\item{v}{Numeric;}

\item{zlev}{Numeric;}

\item{at_zlev}{Numeric;}

\item{zdiff}{Numeric;}

\item{ascending}{Logical;}

\item{dataframe}{Logical;}
}
\description{
\code{\link{wind_shear}} Creates raster from nc path
}
\examples{
\dontrun{
#dont run
}
}
