% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtractor.R
\name{xtractor}
\alias{xtractor}
\title{extractor from atmospheric models}
\usage{
xtractor(x, points, stations, start = "2016-04-15 00:00")
}
\arguments{
\item{x}{raster (or brick)}

\item{points}{data.frame, matrix, SpatialPointsDataFrame.}

\item{stations}{Character; names of stations for each point.}

\item{xtimes}{Logical; names of x are times in format}
}
\value{
Return long data.frame (for ggplot2)
}
\description{
Return data.framee
}
\note{
Based on NCL scrip xtractor from DCA/IAG/USP
}
\examples{
\dontrun{
}
}
