% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_wrf.R
\name{raster_wrf}
\alias{raster_wrf}
\title{Creates raster from nc path}
\usage{
raster_wrf(nc, xmin = -150, xmax = 10, ymin = -60, ymax = 40,
  dx = 215, dy = 135, var, z = 6, array = FALSE, verbose = TRUE)
}
\arguments{
\item{nc}{Character; Path to NetCDF.}

\item{xmin}{Numeric; xmin coordinate.}

\item{xmax}{Numeric; xmax coordinate.}

\item{ymin}{Numeric; ymin coordinate.}

\item{ymax}{Numeric; ymax coordinate.}

\item{dx}{Numeric; grid spacing x.}

\item{dy}{Numeric; grid spacing y.}

\item{var}{Character; name of variable. If you dont know, you can choose.}

\item{z}{Integer; Level (asusming third dimencion)}

\item{array}{Logical; return array or not.}

\item{verbose}{Logical; To show operations.}
}
\value{
raster
}
\description{
\code{\link{raster_nc}} Creates raster from nc path
}
\examples{
\dontrun{
#dont run
}
}
