% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_shear.R
\name{wind_shear}
\alias{wind_shear}
\title{Calculates Wind Shear}
\usage{
wind_shear(u, v, zlev, zlev_bottom, zlev_top, list = FALSE)
}
\arguments{
\item{u}{Numeric; u component of wind}

\item{v}{Numeric; v component of wind}

\item{zlev}{Numeric; levels of each wind}

\item{zlev_bottom}{Numeric; bottom nivels to calculate wind shear, from zlev}

\item{zlev_top}{Numeric; top nivels to calculate wind shear, from zlev}

\item{list}{Logical; Do you want a list of data.frames with all values?}
}
\description{
\code{\link{wind_shear}} Creates raster from nc path
}
\examples{
\dontrun{
#dont run
}
}
