% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtractor.R
\name{xtractor}
\alias{xtractor}
\title{extractor from atmospheric models}
\usage{
xtractor(x, points, stations)
}
\arguments{
\item{x}{raster (or brick)}

\item{points}{data.frame, matrix, SpatialPointsDataFrame.}

\item{stations}{Character; names of stations for each point.}
}
\value{
Return long data.frame (for ggplot2)
}
\description{
Return data.framee
}
\note{
Based on NCL scrip xtractor from DCA/IAG/USP
}
\examples{
\dontrun{
}
}
